fix: Use Date.timeIntervalSinceReferenceDate to avoid rounding precision errors
Swift.Date is internally represented as a Double corresponding to the `timeIntervalSinceReferenceDate`.
Whenever calling Date.timeIntervalSince1970 there is some math internal to Foundation that coverts that Double to the appropriate `timeIntervalSince1970` this will incur some floating point rounding errors for certain Dates depending on the initial precision of the date.

To avoid these sort of problems, the conformance of Date to SQLiteDataConvertible is changed to use `timeIntervalSinceReferenceDate` so that any Foundation.Date maintains the same precision it is input into SQLite with.